+++ packages/shared-metrics/src/util/index.js
@@ -1,0 +1,5 @@
+'use strict';
+
+module.exports = exports = {
+  nativeModuleRetry: require('./nativeModuleRetry')
+};